Add metaclass for AbstractForm and validate ...
FormEmailProcessor.

The metaclass is used to add each non-abstract form in a registry in a
similar way as for Pages. It also checks if an form_processing_backend is
defined and if it is it will call its validate_usage method. This way, the
validate_usage method can throw immediately and ImproperlyConfigured
method so the developer would know if he's done something wrong.

The validate_usage method of the FormEmailProcessor has been implemented
this way.