Add basic form submission views
Add two urls: index which lists all different forms that have been defined
and submissions which for a specific form will list all its submissions.

The index lists each Page seperately (so if we have two different
instances of a form they will be listed seperately here), so the
submissions view will need the app_label, model_name and id to find out
all the submissions of a specific page.