fix saving a single tag with multiple words in it, fixes #824
Add a preprocessTag option for tag-it.js in page-editor.js to fix a problem with saving a single tag with multiple words in it when editing a Page. tag-it.js doesn't put a comma after a single tag (neither double quotes it by default) when POSTing to Wagtail so django-taggit treats it as multiple tags. Force double quoting with mentioned preprocessTag option.