Add support for user models with no username
The User forms were the only known sections to fail when using a swapped
User model that did not have a `username` field, using an `email` field
instead for example.

This commit makes the User forms work with `User.USERNAME_FIELD` instead
of hardcoding `"username"`.

Truly strange user models are not yet supported.

The tests do not test this custom user functionality yet, as this is
quite difficult to do. Lots of code does not support dynamically
changing the `AUTH_USER_MODEL` setting for the tests, so only one custom
user model for most of the tests.

Fixes #1345
See #1380