Move wagtailsettings into contrib module
The `wagtailsettings` module is useful enough that it should be included
in the Wagtail contrib section, to make it available to all Wagtail
developers.

All the code has been given a once-over to make sure it is nice and
polished before being copied in. As such, this is not a direct copy of
the `wagtailsettings` module. It should be backwards compatible though,
excepting the new location.

It has been moved to `wagtail.contrib.settings`, following the naming
scheme set out in #1504.

Documentation has been concatenated in to a single page, and added to
the contrib reference section.