Fix bug where built assets may not be packaged
The asset compilation that occurs as part of `python setup.py sdist`
would find all the files to include first, then recompile all the
assets, before finally packaging them all up. This means that if a new
file was created as part of the asset compilation that was not there
when all the files were first searched for, that file would not be
included in the package.

Now, the asset compilation happens before the file system is searched
for assets to include, so all assets should always be included.