Port wagtailredirects to the permission policy mechanism.
A new helper class in wagtailadmin.utils, PermissionPolicyChecker, is added to provide permission check decorators for the admin views. (This is out of scope for the wagtailcore.permissions module, because it depends on wagtailadmin to supply the 'permission denied' response.)

Not currently implemented - disabling the edit links on the index page listing, for users without change permission. To do this efficiently (and within the constraints of Django templating), we'd need to implement a user/instance-specific lookup object, as we do for pages.