Use HalloPlugin media definitions to import JS for our default plugins.
To make this work, we need an explicit 'default features' list so that the relevant HalloPlugin objects
are available even in the default setup (where no 'features' list is supplied). In this case we
special-case HalloRichTextArea to use the global halloPlugins js config rather than building one
properly via construct_plugins_list, to ensure that legacy setups that hack halloPlugins will
continue to work (for now).